Record test run status for tests doing Okahu trace replay
