+++ test_tools/src/monocle_test_tools/span_loader/__init__.py
@@ -1,0 +1,10 @@
+from . import file_span_loader, okahu_span_loader
+from .file_span_loader import JSONSpanLoader
+from .okahu_span_loader import OkahuSpanLoader
+
+__all__ = [
+    "JSONSpanLoader",
+    "OkahuSpanLoader",
+    "file_span_loader",
+    "okahu_span_loader",
+]
