+++ test_tools/src/monocle_test_tools/span_loader/span_loader.py
@@ -8,7 +8,7 @@ For new code, prefer importing directly from:
 - monocle_test_tools.file_span_loader import JSONSpanLoader
 - monocle_test_tools.okahu_span_loader import OkahuSpanLoader
 """
-from monocle_test_tools.file_span_loader import JSONSpanLoader
-from monocle_test_tools.okahu_span_loader import OkahuSpanLoader
+from monocle_test_tools.span_loader.file_span_loader import JSONSpanLoader
+from monocle_test_tools.span_loader.okahu_span_loader import OkahuSpanLoader
 
 __all__ = ["JSONSpanLoader", "OkahuSpanLoader"]
